Fixed Auth
Navbar:
Moved your styling into navbar.component.css
Changed Login/Logout to match styling you had for Captains/Treasure Chest and made it so Login only shows up when a user is not signed in.

Auth-intercepter:
Commented on the code with changes.

auth.services.ts:

Only a currentUser is getting set to localStorage now, since we can get the token off the currentUser
The BehaviorSubject for the user I am setting to the currentUser in localStorage, if one doesn't exist then it will be null like before, this was the issue with once you left the page the user no longer being found even though it existed in localStorage.

In the nav-bar.component.ts file on ngOnInit you were subscribing the the user subject, with the null value, so on the next page load after login, the user was being set to null.  By doing the above, that fixed that problem.